requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-discogs-client==2.8
pytube==15.0.0
requests==2.32.3
six==1.17.0
toml==0.10.2
urllib3==2.4.0
Werkzeug==3.1.3
youtube-dl==2021.12.17
yt-dlp==2025.3.31
